+++ src/data.rs
@@ -3,6 +3,6 @@ pub mod domain;
 pub mod episode;
 pub mod event;
 pub mod filter;
-pub mod indicator;
+pub mod batch_indicator;
 pub mod query;
 pub mod view;
